+++ src/Resources/contao/dca/tl_ls_shop_coupon.php
@@ -57,7 +57,7 @@ $GLOBALS['TL_DCA']['tl_ls_shop_coupon'] = array(
 				'label'               => &$GLOBALS['TL_LANG']['tl_ls_shop_coupon']['delete'],
 				'href'                => 'act=delete',
 				'icon'                => 'delete.gif',
-				'attributes'          => 'onclick="if (!confirm(\'' . $GLOBALS['TL_LANG']['MSC']['deleteConfirm'] . '\')) return false; Backend.getScrollOffset();"'
+				'attributes'          => 'onclick="if (!confirm(\'' . ($GLOBALS['TL_LANG']['MSC']['deleteConfirm'] ?? null) . '\')) return false; Backend.getScrollOffset();"'
 			),
 			'show' => array(
 				'label'               => &$GLOBALS['TL_LANG']['tl_ls_shop_coupon']['show'],